requirements.txt
gradio==5.5.0
numpy==2.1.3
opencv_python==4.9.0.80
opencv_python_headless==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
PytorchWildlife==1.1.0
PytorchWildlife==1.1.0
supervision==0.24.0
torch==2.3.1
transformers==4.43.4
